database split - moved import_movies and devtool bugfix
moved /import_movies from routes into database
fixed a bug in the debug tools with the qa script not running
