+++ devtools/armui.py
@@ -14,8 +14,9 @@ import log
 #  OUTPUT: none
 def stop():
     try:
+        log.info("Going to stop ARMUI - requesting sudo")
         os.system("sudo systemctl stop armui.service")
-        log.info("ARM UI stopped")
+        log.success("ARM UI stopped")
     except Exception as error:
         log.error(f"ARM UI unable to stop - {error}")
 
@@ -25,7 +26,8 @@ def stop():
 #  OUTPUT: none
 def start():
     try:
+        log.info("Going to restart ARMUI - requesting sudo")
         os.system("sudo systemctl start armui.service")
-        log.info("ARM UI started")
+        log.success("ARM UI started")
     except Exception as error:
         log.error(f"ARM UI unable to start - {error}")
